registration cont.

+++ src/common/bitcoinNodeMedium.h
@@ -135,7 +135,7 @@ CBitcoinNodeMedium< _Type >::reloadResponses()
 {
 	m_responses.clear();
 
-	m_responses.insert( std::make_pair( LastRequest, CRequestedMerkles( m_merkles, m_transactions,reinterpret_cast< long long >( m_node ) ) ) );
+	m_responses.insert( std::make_pair( LastRequest, CRequestedMerkles( m_merkles, m_transactions,reinterpret_cast< uintptr_t >( m_node ) ) ) );
 }
 
 template < class _Type >
